herelaw/requirements.txt
python-dotenv==0.21.1
flask==2.2.3
boto3==1.26.107
pymysql==1.0.3
openai
Pillow
pandas
openpyxl
haversine
bcrypt
PyJWT
pytz
folium
pyppeteer
selenium